Making tests work p2

+++ src/main/java/org/bioinfo/ngs/qc/qualimap/beans/BamStats.java
@@ -2854,7 +2854,7 @@ public class BamStats implements Serializable {
         this.numSelectedRegions = numSelectedRegions;
     }
 
-    public long getNumberOfMappedReadsInRegions() {
+    public int getNumberOfMappedReadsInRegions() {
         return numberOfMappedReadsInRegions;
     }
